Migrate features/content/webinarPercent to TypeScript
- closest('a') from observer target instead of document.querySelector
- match(/lessons\/(\d+)/)?.[1] for proper lesson id extraction
- optional chaining on lastChild walk + instanceof Element guard
- inline tiny getWebinarLinkElement and getWebinarId helpers
