+++ src/chrome-scripts/features/content/index.js
@@ -1,5 +1,5 @@
 export { createHomeworkObserver as homeworkPercent } from './homeworkPercent';
-export { default as webinarPercent } from './webinarPercent.js';
+export { createWebinarObserver as webinarPercent } from './webinarPercent';
 export { createReadingTimeObserver as readingTime } from './readingTime';
 export { default as readingList } from './readingList.js';
 export { default as searchButton } from './searchButton.js';
